fix: cleanup the path and comment
