+++ src/index.ts
@@ -1,2 +1,2 @@
+export { FFIResult, BuildOptions, TestOptions } from 'types'
 export * from './build'
-export { FFIResult, BuildOptions, TestOptions } from './types/index'
