+++ src/lib/index.ts
@@ -1,0 +1,2 @@
+export * from './utils'
+export * from './vm'
