+++ test/contract/dummy/Move.toml
@@ -8,3 +8,4 @@ version = "1.0.0"
 std = "0x1"
 initia_std = "0x1"
 test="_"
+test2 = "_"
