+++ test/contract/dummy/doc/dummy.md
@@ -1,0 +1,62 @@
+
+<a id="0x999_dummy"></a>
+
+# Module `0x999::dummy`
+
+
+
+-  [Function `return_0`](#0x999_dummy_return_0)
+-  [Function `return_10`](#0x999_dummy_return_10)
+
+
+<pre><code></code></pre>
+
+
+
+<a id="0x999_dummy_return_0"></a>
+
+## Function `return_0`
+
+
+
+<pre><code><b>public</b> <b>fun</b> <a href="dummy.md#0x999_dummy_return_0">return_0</a>(): u32
+</code></pre>
+
+
+
+<details>
+<summary>Implementation</summary>
+
+
+<pre><code><b>public</b> <b>fun</b> <a href="dummy.md#0x999_dummy_return_0">return_0</a>(): u32 {
+    0
+}
+</code></pre>
+
+
+
+</details>
+
+<a id="0x999_dummy_return_10"></a>
+
+## Function `return_10`
+
+
+
+<pre><code><b>public</b> <b>fun</b> <a href="dummy.md#0x999_dummy_return_10">return_10</a>(): u32
+</code></pre>
+
+
+
+<details>
+<summary>Implementation</summary>
+
+
+<pre><code><b>public</b> <b>fun</b> <a href="dummy.md#0x999_dummy_return_10">return_10</a>(): u32 {
+    10
+}
+</code></pre>
+
+
+
+</details>
