+++ test/contract/new/Move.toml
@@ -3,7 +3,8 @@ name = "new"
 version = "0.0.0"
 
 [dependencies]
-InitiaStdlib = { git = "https://github.com/initia-labs/movevm.git", subdir = "precompile/modules/initia_stdlib", rev = "main" }
+InitiaStdlib = { git = "https://github.com/initia-labs/move-natives.git", subdir = "initia_stdlib", rev = "main"}
 
 [addresses]
-std =  "0x1"
+std = "0x1"
+
