+++ test/script.spec.ts
@@ -1,6 +1,6 @@
-import { MoveBuilder } from '../src/build'
 import path from 'path'
 import { readFile } from 'fs/promises'
+import { MoveBuilder } from 'build'
 
 describe('build script and decode', () => {
   const contractDir = path.resolve(__dirname, 'contract/script')
